requirements.txt
Flask
mysql-connector-python
python-dotenv
apscheduler
requests
reportlab
pytz
numpy
numpy-financial
